% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_site.R
\name{annotate_site}
\alias{annotate_site}
\title{Annotate site.}
\usage{
annotate_site(site)
}
\arguments{
\item{site}{A data frame from get_site()}
}
\value{
A data frame
}
\description{
Annotate site.
}
