% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{junk_phrases}
\alias{junk_phrases}
\title{A list of non-content terms/phrases.}
\format{
A character vector
}
\usage{
junk_phrases
}
\description{
A list of terms/phrases indicative of comments/advertising, and not article content.
}
\keyword{datasets}
