% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnews_get_gnewsmeta.R
\name{qnews_get_newsmeta}
\alias{qnews_get_newsmeta}
\title{Get URLs/metadata for articles aggregated on Google News per search}
\usage{
qnews_get_newsmeta(term = NULL)
}
\arguments{
\item{term}{Search term}
}
\value{
A data frame
}
\description{
Get URLs/metadata for articles aggregated on Google News per search
}
