% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnews_unshorten_urls.R
\name{qnews_unshorten_urls}
\alias{qnews_unshorten_urls}
\title{Unshorten urls -- mostly from Twitter.}
\usage{
qnews_unshorten_urls(x)
}
\arguments{
\item{x}{A vector of shortened URLs}
}
\value{
A data frame
}
\description{
Unshorten urls -- mostly from Twitter.
}
