% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_web.R
\name{scrape_web}
\alias{scrape_web}
\alias{qnews_scrape_web}
\title{Scrape web articles based on metadata obtained via qnews_get_meta().}
\usage{
qnews_scrape_web(y)
}
\arguments{
\item{link_var}{Variable containing web addresses.}
}
\value{
A corpus data frame object.
}
\description{
This function scrapes news articles from web based on user specified set of web addresses.
}
