% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twt_clean_urls.R
\name{twt_clean_urls}
\alias{twt_clean_urls}
\title{Clean, make uniform URL format -- mostly from Twitter. Identify URLs that have been shortened.}
\usage{
twt_clean_urls(url)
}
\arguments{
\item{url}{A vector of URLs}
}
\value{
A data frame
}
\description{
Clean, make uniform URL format -- mostly from Twitter. Identify URLs that have been shortened.
}
