% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnews_extract_article.R
\name{qnews_extract_article}
\alias{qnews_extract_article}
\title{Extract article content from online news sources.}
\usage{
qnews_extract_article(url, cores)
}
\arguments{
\item{url}{A vector of URLs}

\item{cores}{An integer value specifying n threads}
}
\value{
A data frame
}
\description{
Extract article content from online news sources.
}
